32-bit codegen (#876)

+++ asterius/src/Asterius/Internals/MagicNumber.hs
@@ -9,15 +9,15 @@ where
 import Data.Int
 import Data.Word
 
-invalidAddress :: Int64
+invalidAddress :: Int32
 invalidAddress = 0x001fffffffff0000
 
 -- | Base address for functions. NOTE: reserve 0 for the null function pointer.
 defaultTableBase :: Word32
 defaultTableBase = 1024
 
-mkStaticDataAddress :: Word32 -> Word32 -> Int64
+mkStaticDataAddress :: Word32 -> Word32 -> Int32
 mkStaticDataAddress memory_base off = fromIntegral (memory_base + off)
 
-mkStaticFunctionAddress :: Word32 -> Int64
+mkStaticFunctionAddress :: Word32 -> Int32
 mkStaticFunctionAddress off = fromIntegral (defaultTableBase + off)
